requirements.txt
firebase-admin~=6.5.0
python-dotenv
streamlit